backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-gis
psycopg2-binary
django-cors-headers
gunicorn
GDAL==3.6.2
djangorestframework-gis
